+++ cloudseed/__init__.py
@@ -1,3 +1,3 @@
 """cloudseed: cloud-init VM template customization CLI."""
 
-__version__ = "1.5.2"
+__version__ = "1.5.3"